+++ src/data_types/index.ts
@@ -1,11 +1,5 @@
 export * from './expression';
-export * from '../survey/survey';
-export * from '../survey/survey-file-schema';
-export * from '../survey/items/survey-item';
-export * from '../survey/components/survey-item-component';
-export * from './item-component-key';
 export * from './context';
 export * from './response';
 export * from './utils';
 export * from './legacy-types';
-export * from '../survey/utils/content';
