+++ src/survey/components/survey-item-component.ts
@@ -1,5 +1,5 @@
 import { Expression } from "../../data_types/expression";
-import { ItemComponentKey } from "../../data_types/item-component-key";
+import { ItemComponentKey } from "../item-component-key";
 import { JsonItemComponent } from "../survey-file-schema";
